third_party/WebKit/LayoutTests/shadow-dom/css-style-inherit-expected.txt

A changed inherited property on a slot parent should propagate down to slotted elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(s1).color is "rgb(0, 0, 0)"
PASS getComputedStyle(s1).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
This text should be green
